pyproject.toml
[project]
name = "luiseduromp-rag"
version = "0.1.0"
description = "A personal RAG (Retrieval-Augmented Generation) application to answer questions about my personal profile"
authors = [{ name = "Luis Romero", email = "luiseduromp@hotmail.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain-chroma (>=0.2.4,<0.3.0)",
    "requests (>=2.32.4,<3.0.0)",
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "app" }]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.20.0"
black = "^23.0.0"
ruff = "^0.1.0"
pre-commit = "^3.0.0"
boto3 = "^1.39.3"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py312"
fix = true
ignore = []

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.black]
line-length = 88
target-version = ["py312"]